#28 Organized Model Classes into packages
moved model classes into separate packages
