+++ junit/edu/ycp/cs320/assign01/model/LibraryTest.java
@@ -8,6 +8,10 @@ import java.util.ArrayList;
 import org.junit.Before;
 import org.junit.Test;
 
+import edu.ycp.cs320.assign01.model.movement.Location;
+import edu.ycp.cs320.assign01.model.movement.Room;
+import edu.ycp.cs320.assign01.model.utility.Library;
+
 public class LibraryTest {
 	private Library lib;
