+++ src/edu/ycp/cs320/assign01/model/NPC.java
@@ -3,6 +3,10 @@ package edu.ycp.cs320.assign01.model;
 import java.util.ArrayList;
 import java.util.Random;
 
+import edu.ycp.cs320.assign01.model.interfaces.Named;
+import edu.ycp.cs320.assign01.model.utility.Pair;
+import edu.ycp.cs320.assign01.model.utility.Triple;
+
 public class NPC implements Named {
 	private String name, shortDesc, longDesc;
 	private int level, health, id, minAttack, maxAttack;
