+++ src/edu/ycp/cs320/assign01/model/interfaces/Named.java
@@ -1,4 +1,4 @@
-package edu.ycp.cs320.assign01.model;
+package edu.ycp.cs320.assign01.model.interfaces;
 
 public interface Named {
 	public void setName(String name);
