+++ src/edu/ycp/cs320/assign01/model/movement/Location.java
@@ -1,6 +1,9 @@
-package edu.ycp.cs320.assign01.model;
+package edu.ycp.cs320.assign01.model.movement;
 
 import java.util.ArrayList;
+
+import edu.ycp.cs320.assign01.model.interfaces.Named;
+import edu.ycp.cs320.assign01.model.interfaces.Navigable;
 
 public class Location implements Navigable, Named {
 	private ArrayList<Room> roomList;
